retrieve_summary/requirements.txt
boto3
python-telegram-bot
requests
